chore: add AGENTS.md, makefile, update README, add golangci and gofumpt

+++ gui_small_logo.go
@@ -2,14 +2,16 @@ package main
 
 import "github.com/veandco/go-sdl2/sdl"
 
+// smallLogoCrop is the sprite crop for the game logo.
 var smallLogoCrop = sdl.Rect{X: 700, Y: 401, W: 115, H: 58}
 
+// RenderSmallLogo renders the game logo in the top-right corner.
 func RenderSmallLogo() {
 	dest := sdl.Rect{
-		X: int32(WindowW - smallLogoCrop.W),
+		X: WindowW - smallLogoCrop.W,
 		Y: 0,
-		W: int32(smallLogoCrop.W),
-		H: int32(smallLogoCrop.H),
+		W: smallLogoCrop.W,
+		H: smallLogoCrop.H,
 	}
 
 	RenderSprite(TextureMain, dest, smallLogoCrop)
